Organize and add a lot of documentation for layouts
