+++ .swiftlint.yml
@@ -7,7 +7,9 @@ disabled_rules:
 included:
     - Amethyst
     - AmethystTests
-line_length: 200
+line_length:
+    warning: 200
+    ignores_comments: true
 file_length: 500
 type_body_length: 400
 cyclomatic_complexity: 15
