+++ Amethyst/Layout/Layouts/FloatingLayout.swift
@@ -14,7 +14,7 @@ final class FloatingLayout<Window: WindowType>: Layout<Window> {
 
     override var layoutDescription: String { return "" }
 
-    override func reflow(_ windows: [AnyWindow<Window>], on screen: NSScreen) -> Operation? {
+    override func reflow(_ windows: [AnyWindow<Window>], on screen: NSScreen) -> ReflowOperation<Window>? {
         return nil
     }
 }
