+++ Amethyst/Layout/Layouts/WideLayout.swift
@@ -82,7 +82,7 @@ final class WideLayout<Window: WindowType>: Layout<Window>, PanedLayout {
         mainPaneCount = max(1, mainPaneCount - 1)
     }
 
-    override func reflow(_ windows: [AnyWindow<Window>], on screen: NSScreen) -> Operation? {
+    override func reflow(_ windows: [AnyWindow<Window>], on screen: NSScreen) -> ReflowOperation<Window>? {
         let assigner = Assigner(windowActivityCache: windowActivityCache)
         return WideReflowOperation(screen: screen, windows: windows, layout: self, frameAssigner: assigner)
     }
