+++ Amethyst/Layout/Layouts/WidescreenTallLayout.swift
@@ -72,7 +72,7 @@ final class WidescreenTallLayout<Window: WindowType>: Layout<Window> {
     private(set) var mainPaneCount: Int = 1
     private(set) var mainPaneRatio: CGFloat = 0.5
 
-    override func reflow(_ windows: [AnyWindow<Window>], on screen: NSScreen) -> Operation? {
+    override func reflow(_ windows: [AnyWindow<Window>], on screen: NSScreen) -> ReflowOperation<Window>? {
         let assigner = Assigner(windowActivityCache: windowActivityCache)
         return WidescreenTallReflowOperation(screen: screen, windows: windows, layout: self, frameAssigner: assigner)
     }
