+++ Amethyst/Model/WindowsInformation.swift
